requirements.txt
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
crcmod==1.7
Cython==0.29.6
dateutils==0.6.6
dbus-python==1.2.8
Flask==1.0.3
future==0.17.1
gpsd-py3==0.3.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
numpy==1.17.0.dev0+d7a73f8
pandas==0.24.2
Pillow==6.0.0
pycparser==2.19
pyFFTW==0.11.1
pyiridium9602==1.2.3
PyPRUSS==0.1
pyserial==3.4
python-dateutil==2.8.0
pytz==2018.9
pywt==1.0.6
PyYAML==5.1
requests==2.21.0
schedule==0.6.0
scikit-learn==0.20.3
scipy==1.2.1
scons==3.0.5
serial==0.0.97
six==1.12.0
smbus==1.1.post2
SoundFile==0.10.2
urllib3==1.24.1
Werkzeug==0.15.4
